调整代码和结构
