+++ bot.py
@@ -1,0 +1,9 @@
+from os import path
+import none
+
+import config
+
+if __name__ == '__main__':
+    none.init(config)
+    none.load_plugins(path.join(path.dirname(__file__), 'plugins'), 'plugins')
+    none.run()
