+++ config.py
@@ -1,0 +1,7 @@
+from none.default_config import *
+
+HOST = '0.0.0.0'
+PORT = 8080
+
+SUPERUSERS = {1320163325, 1206985125}
+COMMAND_START.add('')
